web/requirements.txt
Django>=2.0
psycopg2
celery
django-allauth
djangorestframework<=3.8.2
Pillow==2.8.2
mock
django-nose
